+++ src/Middleware/CallableMiddlewareWrapper.php
@@ -2,9 +2,9 @@
 
 namespace Rougin\Slytherin\Middleware;
 
-use Interop\Http\ServerMiddleware\DelegateInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use Rougin\Slytherin\Middleware\HandlerInterface;
 
 /**
  * Callable Middleware Wrapper
@@ -43,22 +43,22 @@ class CallableMiddlewareWrapper implements MiddlewareInterface
     /**
      * Processes an incoming server request and return a response.
      *
-     * @param  \Psr\Http\Message\ServerRequestInterface         $request
-     * @param  \Interop\Http\ServerMiddleware\DelegateInterface $delegate
+     * @param  \Psr\Http\Message\ServerRequestInterface      $request
+     * @param  \Rougin\Slytherin\Middleware\HandlerInterface $handler
      * @return \Psr\Http\Message\ResponseInterface
      */
-    public function process(ServerRequestInterface $request, DelegateInterface $delegate)
+    public function process(ServerRequestInterface $request, HandlerInterface $handler)
     {
         $middleware = $this->middleware;
 
         if ($this->response instanceof ResponseInterface) {
-            $delegate = function ($request) use ($delegate) {
-                return $delegate->process($request);
+            $handler = function ($request) use ($handler) {
+                return $handler->{HANDLER_METHOD}($request);
             };
 
-            return $middleware($request, $this->response, $delegate);
+            return $middleware($request, $this->response, $handler);
         }
 
-        return $middleware($request, $delegate);
+        return $middleware($request, $handler);
     }
 }