+++ tests/Application/ApplicationTestCases.php
@@ -89,10 +89,6 @@ class ApplicationTestCases extends \PHPUnit_Framework_TestCase
      */
     public function testHandleMethodWithMiddleware()
     {
-        $interface = 'Interop\Http\ServerMiddleware\MiddlewareInterface';
-
-        interface_exists($interface) || $this->markTestSkipped('PSR-15 is not installed.');
-
         $request = $this->request('GET', '/middleware');
 
         $expected = (string) 'Loaded with middleware';